TASK: Extract creating of content repository nodes to a trait
