+++ Tests/Functional/Eel/ElasticSearchQueryTest.php
@@ -13,34 +13,20 @@ namespace Flowpack\ElasticSearch\ContentRepositoryAdaptor\Tests\Functional\Eel;
  * source code.
  */
 
-use DateTime;
 use DateTimeImmutable;
 use Exception;
 use Flowpack\ElasticSearch\ContentRepositoryAdaptor\Command\NodeIndexCommandController;
 use Flowpack\ElasticSearch\ContentRepositoryAdaptor\Eel\ElasticSearchQueryBuilder;
 use Flowpack\ElasticSearch\ContentRepositoryAdaptor\Eel\ElasticSearchQueryResult;
 use Flowpack\ElasticSearch\ContentRepositoryAdaptor\Exception\QueryBuildingException;
-use Flowpack\ElasticSearch\Transfer\Exception\ApiException;
+use Flowpack\ElasticSearch\ContentRepositoryAdaptor\Tests\Functional\TRaits\ContentRepositoryNodeCreationTrait;
 use Neos\ContentRepository\Domain\Model\NodeInterface;
-use Neos\ContentRepository\Domain\Model\Workspace;
-use Neos\ContentRepository\Domain\Repository\NodeDataRepository;
-use Neos\ContentRepository\Domain\Repository\WorkspaceRepository;
-use Neos\ContentRepository\Domain\Service\Context;
-use Neos\ContentRepository\Domain\Service\ContextFactoryInterface;
-use Neos\ContentRepository\Domain\Service\NodeTypeManager;
-use Neos\ContentRepository\Exception\NodeExistsException;
-use Neos\ContentRepository\Exception\NodeTypeNotFoundException;
-use Neos\Flow\Cli\Exception\StopCommandException;
-use Neos\Flow\Mvc\Exception\StopActionException;
 use Neos\Flow\Persistence\QueryResultInterface;
 use Neos\Flow\Tests\FunctionalTestCase;
 
 class ElasticSearchQueryTest extends FunctionalTestCase
 {
-    /**
-     * @var WorkspaceRepository
-     */
-    protected $workspaceRepository;
+    use ContentRepositoryNodeCreationTrait;
 
     /**
      * @var NodeIndexCommandController
@@ -48,34 +34,9 @@ class ElasticSearchQueryTest extends FunctionalTestCase
     protected $nodeIndexCommandController;
 
     /**
-     * @var Context
-     */
-    protected $context;
-
-    /**
-     * @var NodeInterface
-     */
-    protected $siteNode;
-
-    /**
      * @var boolean
      */
     protected static $testablePersistenceEnabled = true;
-
-    /**
-     * @var ContextFactoryInterface
-     */
-    protected $contextFactory;
-
-    /**
-     * @var NodeTypeManager
-     */
-    protected $nodeTypeManager;
-
-    /**
-     * @var NodeDataRepository
-     */
-    protected $nodeDataRepository;
 
     /**
      * @var boolean
@@ -85,27 +46,13 @@ class ElasticSearchQueryTest extends FunctionalTestCase
     public function setUp(): void
     {
         parent::setUp();
-        $this->workspaceRepository = $this->objectManager->get(WorkspaceRepository::class);
-        $liveWorkspace = new Workspace('live');
-        $this->workspaceRepository->add($liveWorkspace);
-
-        $this->nodeTypeManager = $this->objectManager->get(NodeTypeManager::class);
-        $this->contextFactory = $this->objectManager->get(ContextFactoryInterface::class);
-        $this->context = $this->contextFactory->create([
-            'workspaceName' => 'live',
-            'dimensions' => ['language' => ['en_US']],
-            'targetDimensions' => ['language' => 'en_US']
-        ]);
-        $rootNode = $this->context->getRootNode();
-
-        $this->siteNode = $rootNode->createNode('welcome', $this->nodeTypeManager->getNodeType('Neos.NodeTypes:Page'));
-        $this->siteNode->setProperty('title', 'welcome');
-
-        $this->nodeDataRepository = $this->objectManager->get(NodeDataRepository::class);
+        $this->setupContentRepository();
 
         $this->nodeIndexCommandController = $this->objectManager->get(NodeIndexCommandController::class);
 
         $this->createNodesForNodeSearchTest();
+        sleep(2);
+        $this->indexNodes();
     }
 
     public function tearDown(): void
@@ -394,52 +341,8 @@ class ElasticSearchQueryTest extends FunctionalTestCase
         return substr(strrchr($method, '\\'), 1);
     }
 
-    /**
-     * Creates some sample nodes to run tests against
-     *
-     * @throws NodeExistsException
-     * @throws NodeTypeNotFoundException
-     * @throws StopActionException
-     * @throws \Flowpack\ElasticSearch\ContentRepositoryAdaptor\Exception
-     * @throws ApiException
-     * @throws StopCommandException
-     */
-    protected function createNodesForNodeSearchTest(): void
+    protected function indexNodes(): void
     {
-        $newDocumentNode1 = $this->siteNode->createNode('test-node-1', $this->nodeTypeManager->getNodeType('Neos.NodeTypes:Page'));
-        $newDocumentNode1->setProperty('title', 'chicken');
-        $newDocumentNode1->setProperty('title_analyzed', 'chicken');
-
-        $newContentNode1 = $newDocumentNode1->getNode('main')->createNode('document-1-text-1', $this->nodeTypeManager->getNodeType('Neos.NodeTypes:Text'));
-        $newContentNode1->setProperty('text', 'A Scout smiles and whistles under all circumstances.');
-
-        $newDocumentNode2 = $this->siteNode->createNode('test-node-2', $this->nodeTypeManager->getNodeType('Neos.NodeTypes:Page'));
-        $newDocumentNode2->setProperty('title', 'chicken');
-        $newDocumentNode2->setProperty('title_analyzed', 'chicken');
-
-        // Nodes for cacheLifetime test
-        $newContentNode2 = $newDocumentNode2->getNode('main')->createNode('document-2-text-1', $this->nodeTypeManager->getNodeType('Neos.NodeTypes:Text'));
-        $newContentNode2->setProperty('text', 'Hidden after 2025-01-01');
-        $newContentNode2->setHiddenAfterDateTime(new DateTime('@1735686000'));
-        $newContentNode3 = $newDocumentNode2->getNode('main')->createNode('document-2-text-2', $this->nodeTypeManager->getNodeType('Neos.NodeTypes:Text'));
-        $newContentNode3->setProperty('text', 'Hidden before 2018-07-18');
-        $newContentNode3->setHiddenBeforeDateTime(new DateTime('@1531864800'));
-
-        $newDocumentNode3 = $this->siteNode->createNode('test-node-3', $this->nodeTypeManager->getNodeType('Neos.NodeTypes:Page'));
-        $newDocumentNode3->setProperty('title', 'egg');
-        $newDocumentNode3->setProperty('title_analyzed', 'egg');
-
-        $dimensionContext = $this->contextFactory->create([
-            'workspaceName' => 'live',
-            'dimensions' => ['language' => ['de']]
-        ]);
-        $translatedNode3 = $dimensionContext->adoptNode($newDocumentNode3, true);
-        $translatedNode3->setProperty('title', 'De');
-
-        $this->persistenceManager->persistAll();
-
-        sleep(2);
-
         if (self::$indexInitialized === true) {
             return;
         }
